requirements.txt
flask
pyyaml
requests
paho-mqtt
Pillow
numpy
tflite-support
